Removed usage of stats service. Moved code to pipe
