+++ src/app/app.module.ts
@@ -37,6 +37,7 @@ import { NotFoundPageComponent } from './pages/not-found-page/not-found-page.com
 import { SetNameComponent } from './pages/set-name/set-name.component';
 import { StatsPageComponent } from './pages/stats-page/stats-page.component';
 import { XchPipe } from './pipes/xch.pipe';
+import { MinutesPipe } from './pipes/minutes.pipe'
 import { PoolSizesEffects } from './stores/pool-sizes/pool-sizes.effects';
 import { poolSizesReducer } from './stores/pool-sizes/pool-sizes.reducer';
 import { poolSizeStateKey } from './stores/pool-sizes/pool-sizes.selectors';
@@ -74,6 +75,7 @@ import { poolSizeStateKey } from './stores/pool-sizes/pool-sizes.selectors';
 
     // Pipes
     XchPipe,
+    MinutesPipe,
   ],
   imports: [
     BrowserModule,
