+++ src/app/pages/stats-page/payouts.store.ts
@@ -2,12 +2,8 @@ import { Injectable } from '@angular/core';
 import { ComponentStore } from '@ngrx/component-store';
 import { EMPTY, Observable } from 'rxjs';
 import { catchError, switchMap, takeUntil, tap } from 'rxjs/operators';
-import { CoinRecord } from '../../interfaces/coin-record.interface';
 import { Payout, PayoutAddress } from '../../interfaces/payout.interface';
-import { PoolStats } from '../../interfaces/pool-stats.interface';
-import { CoinRecordService } from '../../services/api/coin-record.service';
 import { PayoutService } from '../../services/api/payout.service';
-import { PoolService } from '../../services/api/pool.service';
 
 export interface PayoutsState {
   isLoading: boolean;
