+++ src/app/pages/stats-page/stats-page.component.html
@@ -12,7 +12,7 @@
         <div class="big-number-title">Last Reward</div>
       </div>
       <div class="big-number">
-        <div class="big-number-value even-longer-value">{{ format_minutes(info.minutes_to_win) }}</div>
+        <div class="big-number-value even-longer-value">{{ info.minutes_to_win | minutes }}</div>
         <div class="big-number-title">Estimated Time To Win</div>
       </div>
       <div class="break"></div>
